+++ fydp-web/src/main/java/com/thomasdedinsky/fydp/fydpweb/data/HubLocationRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface HubLocationRepository extends JpaRepository<HubLocation, Integer> {
+public interface HubLocationRepository extends JpaRepository<HubLocation, Long> {
     List<HubLocation> findByHub(Hub hub);
     Page<HubLocation> findByHub(Hub hub, Pageable pageable);
     List<HubLocation> findByIdIn(List<Long> idList);
